Rename imageservice -> imageService
