+++ packages/11ty/_includes/components/figure/image-service.js
